created rest controller, moved methods that return response body to new controller, working
