+++ mgl-webapp/src/main/java/com/organization/mvcproject/controller/GameController.java
@@ -3,7 +3,6 @@ package com.organization.mvcproject.controller;
 
 import java.util.List;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -54,20 +53,6 @@ public class GameController {
 		return "gamesPage";
 	}
 
-	/**
-	 * TODO 2.0 (Separation of concerns) consider moving all controller endpoints that return a ResponseEntity into a @RestController.
-	 */
 	
 
-	@GetMapping(value="/game")
-	public ResponseEntity<List<Game>> fetchAllGames() {
-		return new ResponseEntity<List<Game>>(gameService.retrieveAllGames(), HttpStatus.OK);
-	}
-
-
-	@RequestMapping(value = "/game", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Void> createGame(@RequestBody Game game) {
-		gameService.saveGame(game);
-		return new ResponseEntity<Void>(HttpStatus.CREATED);
-	}
 }
