+++ mgl-webapp/src/main/java/com/organization/mvcproject/controller/GameRestController.java
@@ -1,0 +1,36 @@
+package com.organization.mvcproject.controller;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.organization.mvcproject.model.Game;
+import com.organization.mvcproject.service.GameService;
+
+@RestController
+public class GameRestController {
+
+
+	@Autowired
+	private GameService gameService;
+
+	@GetMapping(value="/game")
+	public ResponseEntity<List<Game>> fetchAllGames() {
+		return new ResponseEntity<List<Game>>(gameService.retrieveAllGames(), HttpStatus.OK);
+	}
+
+
+	@RequestMapping(value = "/game", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity<Void> createGame(@RequestBody Game game) {
+		gameService.saveGame(game);
+		return new ResponseEntity<Void>(HttpStatus.CREATED);
+	}
+}
